Introducing Reader::fetchPairs
In order to make League\Csv works better with huge
CSV files we introduce two modes where some Reader::fetch*
methods will return either an array or an iterator. The mode
is defined for a single query and is by default reset to returning
an array.

The mode are added with the introduction of Reader::fetchPairs to
returns data in an array of key-value pairs, as an associative array
with a single entry per row.

Adding Reader::fetchPairs requires dropping support for PHP5.4 as the
method requires the use of PHP 5.5+ generators.

The fetch* method callable function are also updated to ease their usage as
prior to Reader::fetchPairs all callable had the same signature. Now the signature
depends on the fetch* method.